Add `CustomSpecStyle` and `SuiteSpec` for enhanced support of custom test styles
