+++ kotest-intellij-plugin/src/main/kotlin/io/kotest/plugin/intellij/styles/SpecStyle.kt
@@ -39,6 +39,7 @@ interface SpecStyle {
          ShouldSpecStyle,
          WordSpecStyle,
          AnnotationSpecStyle,
+         CustomSpecStyle,
       )
 
       /**
